# docs/swagger.yaml
basePath: /
definitions:
  gin-scalable-api_pkg_response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_modules_branch.BranchResponse:
    properties:
      code:
        type: string
      company_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      level:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_modules_company.CompanyListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_modules_company.CompanyResponse'
        type: array
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  internal_modules_company.CompanyResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  internal_modules_company.CreateCompanyRequest:
    properties:
      code:
        maxLength: 20
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - code
    - name
    type: object
  internal_modules_company.UpdateCompanyRequest:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Complete API for ERP with RBAC and API Documentation System. This API provides comprehensive endpoints for managing users, companies, branches, roles, modules, subscriptions, units, applications, and authentication with JWT-based security.

    ## Environments
    - **Development**: http://localhost:8081
    - **Staging**: https://staging-api.huminor.com
    - **Production**: https://api.huminor.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Huminor Console API
  version: "1.0"
paths:
...
  /api/v1/branches/company/{companyId}:
    get:
      consumes:
      - application/json
      description: Mendapatkan semua branches dari company tertentu. Mendukung nested
        hierarchy dengan query parameter nested=true
      parameters:
      - description: Company ID
        in: path
        name: companyId
        required: true
        type: integer
      - description: Include hierarchy information (true/false)
        in: query
        name: include_hierarchy
        type: string
      - description: Return nested hierarchy (true/false)
        in: query
        name: nested
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Company branches berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_modules_branch.BranchResponse'
                  type: array
              type: object
        "400":
          description: Bad request - Invalid company ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Company tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get company branches
      tags:
      - Branches
  /api/v1/companies:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar semua company dengan filter opsional dan pagination
      parameters:
      - description: Limit jumlah data
        in: query
        name: limit
        type: integer
      - description: Offset data
        in: query
        name: offset
        type: integer
      - description: Search by name atau code
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Companies berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyListResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get all companies
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Membuat company baru
      parameters:
      - description: Company data
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/internal_modules_company.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Company berhasil dibuat
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyResponse'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "409":
          description: Conflict - company code sudah ada
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create new company
      tags:
      - Companies
  /api/v1/companies/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus company berdasarkan ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Company berhasil dihapus
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "400":
          description: Bad request - Invalid company ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Company tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete company
      tags:
      - Companies
    get:
      consumes:
      - application/json
      description: Mendapatkan detail company berdasarkan ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Company berhasil diambil
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyResponse'
              type: object
        "400":
          description: Bad request - Invalid company ID
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Company tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get company by ID
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Memperbarui informasi company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company data yang akan diupdate
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/internal_modules_company.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Company berhasil diupdate
          schema:
            allOf:
            - $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_company.CompanyResponse'
              type: object
        "400":
          description: Bad request - Invalid company ID atau validation failed
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "404":
          description: Company tidak ditemukan
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gin-scalable-api_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update company
      tags:
      - Companies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...